requirements.txt
atproto==0.0.46
matplotlib==3.9.0
numpy==1.26.4
pandas==2.2.2
pylint==3.1.0
python-dotenv==1.0.1
python_dateutil==2.9.0.post0
Requests==2.32.2
SQLAlchemy==2.0.30
